factor out common code
